refactor: pause/resume immediately after player update
Before, we would wait until the position of the moved player would be transmitted to the other
users, calculate the distance at that client, and then decide to pause or resume. During this time
the moving player could have disconnected. Instead, this PR immediately sets the consumer of the
moving player AND all receiving players to pause at the same time. This hopefully prevents some
timing issues.