chore: fix clean release.
The Dockerfile has an "optional" COPY statement that copies over generated development SSL certificates. However COPY fails when given no inputs. So on a clean repo it currently doesn't work. I'm removing the COPY statement from the Dockerfile and instead using (another) volume mount for that. Unfortunately that does require the certificate files to live in a subdirectory.